# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LAB APP - API Documentation",
        "description": "API para gestão de escolas.",
        "contact": {
            "email": "pedro.manuel.dev@gmail.com"
        },
        "license": {
            "name": "License Name",
            "url": "http://url-to-license.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/school": {
            "get": {
                "tags": [
                    "Schools"
                ],
                "summary": "Obter uma lista com todas as escolas.",
                "operationId": "getAllSchools",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Uma mensagem de sucesso",
                                            "type": "string",
                                            "example": "Listagem de todas as escolas"
                                        },
                                        "data": {
                                            "description": "Um array com todas as escolas encontradas",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "ID da escola",
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "description": "Nome da escola",
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "description": "E-mail da escola",
                                                        "type": "string"
                                                    },
                                                    "classrooms": {
                                                        "description": "Número de salas da escola",
                                                        "type": "integer"
                                                    },
                                                    "province_id": {
                                                        "description": "ID da provincia",
                                                        "type": "integer"
                                                    },
                                                    "province": {
                                                        "description": "Provincia onde esta localizada a escola.",
                                                        "properties": {
                                                            "id": {
                                                                "description": "ID da provincia",
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "description": "Nome da provincia",
                                                                "type": "string"
                                                            },
                                                            "description": {
                                                                "description": "Descrição da provincia",
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Schools"
                ],
                "summary": "Criar uma nova escola",
                "operationId": "storeSchool",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "classrooms",
                                    "province"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome da escola (Max. 255 characteres)",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "E-mail da escola",
                                        "type": "string"
                                    },
                                    "classrooms": {
                                        "description": "Numero de salas da escola",
                                        "type": "integer"
                                    },
                                    "province_id": {
                                        "description": "ID da provincia onde a escola esta localizada ",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Erro no formulário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Mensagem de erros",
                                            "type": "string",
                                            "example": "Erro no preechimento do formulário"
                                        },
                                        "errors": {
                                            "description": "Uma lista detalhada dos erros encontrado no formulário.",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Uma mensagem de sucesso",
                                            "type": "string",
                                            "example": "Escola criada com sucesso"
                                        },
                                        "data": {
                                            "description": "Retorna a escola criada",
                                            "properties": {
                                                "id": {
                                                    "description": "ID da escola",
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "description": "Nome da escola",
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "description": "O E-mail da escola",
                                                    "type": "string"
                                                },
                                                "classrooms": {
                                                    "description": "Número de salas da escola",
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/school-excel": {
            "post": {
                "tags": [
                    "Schools"
                ],
                "summary": "Importar dados das escolas apartir de um arquivo excel",
                "operationId": "importSchoolsFromExcel",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "excel"
                                ],
                                "properties": {
                                    "excel": {
                                        "description": "Um arquivo excel com dados das ecolas  (Deve ser um arquivo .xlsx)",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Erro no formúlario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Uma mensagem de erro",
                                            "type": "string"
                                        },
                                        "errors": {
                                            "description": "Uma lista detalhada dos erros encontrados no formulário",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Uma mensagem de sucesso",
                                            "type": "string",
                                            "example": "Escolas importadas com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}